+++ c8y_api/model/alarms.py
@@ -334,7 +334,7 @@ class Alarms(CumulocityResource):
             page_number,
             limit,
             Alarm.from_json if not as_tuples else
-            lambda x: parse_as_tuples(x, *([as_tuples] if isinstance(as_tuples, str) else as_tuples)))
+            lambda x: parse_as_tuples(x, as_tuples))
 
     def get_all(
             self,