+++ c8y_api/model/audit.py
@@ -8,7 +8,7 @@ from typing import Generator, List, ClassVar
 
 from c8y_api._base_api import CumulocityRestApi
 from c8y_api.model._base import CumulocityResource, ComplexObject
-from c8y_api.model._parser import ComplexObjectParser, SimpleObjectParser
+from c8y_api.model._parser import ComplexObjectParser, SimpleObjectParser, as_tuples as parse_as_tuples
 from c8y_api.model._util import _DateUtil
 
 
@@ -207,6 +207,7 @@ class AuditRecords(CumulocityResource):
             min_age: timedelta = None, max_age: timedelta = None,
             reverse: bool = False, limit: int = None,
             page_size: int = 1000, page_number: int = None,
+            as_tuples: str | tuple | list[str | tuple] = None,
             **kwargs
     ) -> Generator[AuditRecord]:
         """Query the database for audit records and iterate over the results.
@@ -239,6 +240,10 @@ class AuditRecords(CumulocityResource):
                 parsed in one chunk). This is a performance related setting.
             page_number (int): Pull a specific page; this effectively disables
                 automatic follow-up page retrieval.
+            as_tuples: (*str|tuple):  Don't parse objects, but directly extract
+                the values at certain JSON paths as tuples; If the path is not
+                defined in a result, None is used; Specify a tuple to define
+                a proper default value for each path.
 
         Returns:
             Generator for AuditRecord objects
@@ -250,7 +255,12 @@ class AuditRecords(CumulocityResource):
             min_age=min_age, max_age=max_age,
             reverse=reverse, page_size=page_size,
             **kwargs)
-        return super()._iterate(base_query, page_number, limit, AuditRecord.from_json)
+        return super()._iterate(
+            base_query,
+            page_number,
+            limit,
+            AuditRecord.from_json if not as_tuples else
+            lambda x: parse_as_tuples(x, as_tuples))
 
     def get_all(
             self,
@@ -260,6 +270,7 @@ class AuditRecords(CumulocityResource):
             min_age: timedelta = None, max_age: timedelta = None,
             reverse: bool = False, limit: int = None,
             page_size: int = 1000, page_number: int = None,
+            as_tuples: str | tuple | list[str | tuple] = None,
             **kwargs
     ) -> List[AuditRecord]:
         """Query the database for audit records and return the results as list.
@@ -278,6 +289,7 @@ class AuditRecords(CumulocityResource):
             before=before, after=after,
             min_age=min_age, max_age=max_age,
             reverse=reverse, limit=limit, page_size=page_size, page_number=page_number,
+            as_tuples=as_tuples,
             **kwargs,
         ))