+++ c8y_api/model/events.py
@@ -345,7 +345,7 @@ class Events(CumulocityResource):
             page_number,
             limit,
             Event.from_json if not as_tuples else
-            lambda x: parse_as_tuples(x, *([as_tuples] if isinstance(as_tuples, str) else as_tuples)))
+            lambda x: parse_as_tuples(x, as_tuples))
 
     def get_all(
             self,